+++ action_server/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- A new argument: `--parent-pid` can be passed to the action server. When passed,
+  the action server will automatically exit when the given parent pid is no longer
+  alive (it's meant to be used when the action server is embedded into another app).
+
 ## 0.12.1 - 2024-05-31
 
 - Fix issue where keyring backend was not loaded properly with Windows and MacOS